requirements.txt
absl-py==0.12.0
astunparse==1.6.3
beautifulsoup4==4.6.0
cachetools==4.2.2
certifi==2020.12.5
chardet==3.0.4
click==8.0.0
colorama==0.4.4
filelock==3.0.12
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googletrans==3.1.0a0
grpcio==1.32.0
h11==0.9.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.0.8
hyperframe==5.2.0
idna==2.10
joblib==1.0.1
JPype1==1.2.1
Keras-Preprocessing==1.1.2
kobert-transformers==0.4.1
konlpy==0.5.2
lxml==4.6.3
Markdown==3.3.4
mecab-python===0.996-ko-0.9.2
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
Pillow==8.2.0
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pyparsing==2.4.7
PySocks==1.7.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
sacremoses==0.0.45
six==1.15.0
sniffio==1.2.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tokenizers==0.10.2
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
transformers==4.6.0
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==2.0.0
wrapt==1.12.1
xmltodict==0.12.0
